refactor tiles out of plastic
