+++ src/core/common/utils.ts
@@ -1,3 +1,6 @@
+export type GetOrUndefined<T extends object, K extends string | number | symbol> = 
+  T extends T ? (K extends keyof T ? T[K] : undefined) : undefined;
+export type KeyOfUnion<T> = T extends T ? keyof T: never;
 export type Mutable<T> = { -readonly [P in keyof T]: T[P] };
 
 export function pairEach<T>(
@@ -8,7 +11,6 @@ export function pairEach<T>(
     fn(it[i], it[i + 1], i);
   }
 }
-
 export function pairMap<U, V>(
   it: readonly U[],
   fn: (a: U, b: U, i: number) => V,
