+++ src/core/e2e_tests/serialize.test.ts
@@ -27,9 +27,9 @@ import { Erosion, Landslide } from "../models/hazards";
 import { MinerFactory } from "../models/miner";
 import { atCenterOfTile, position } from "../models/position";
 import { Tile } from "../models/tiles";
-import discover from "../transformers/02_plastic/04_discover";
-import fence from "../transformers/02_plastic/08_fence";
-import serialize from "../transformers/02_plastic/10_serialize";
+import discover from "../transformers/03_plastic/04_discover";
+import fence from "../transformers/03_plastic/08_fence";
+import serialize from "../transformers/03_plastic/10_serialize";
 import goldenTest from "./golden";
 
 function fill<T>(
