+++ src/core/transformers/03_plastic/05_populate.ts
