+++ src/core/transformers/03_plastic/07_enscribe.ts
