+++ src/core/transformers/03_plastic/08_fence.ts
